moar enemies

+++ app/src/main/java/ru/newprotech/invaders/CEnemyBullet.java
@@ -6,7 +6,7 @@ import java.lang.ref.WeakReference;
  * Created by 6003 on 21.01.2015.
  */
 public class CEnemyBullet extends CSpritedThinker{
-    static CEnemyBullet createBullet(float x, float y, float v, int a) {
+    static CEnemyBullet createBullet(float x, float y, float v, float a) {
         CEnemyBullet bullet =  new CEnemyBullet(x, y, v, a);
         CEnemyBulletManager bulletManager = CEnemyBulletManager.getInstance();
         bulletManager.Add(bullet);
@@ -27,7 +27,7 @@ public class CEnemyBullet extends CSpritedThinker{
         return result;
     }
 
-    private CEnemyBullet(float x, float y, float v, int a){
+    private CEnemyBullet(float x, float y, float v, float a){
         CSprite newSprite = CSprite.createSprite(R.drawable.lightning, x, y);
         newSprite.setAnimation(500, 0, 1);
         newSprite.setVx((float) (v * Math.cos(Math.toRadians(a + 90))));
